+++ js/data.js
@@ -1,0 +1,15 @@
+const sms = [
+  'Всё отлично!', 'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+// Массив с сообщениями для фотографии
+const names = [
+  'Артем', 'Дмитрий', 'Алексей', 'Егор', 'Лиза', 'Михаил', 'Александр', 'Света'
+];
+
+//массив с именами
+
+export { sms, names };
